Rename tools to menu, because that's where menu items are Put all the DCC interfaces into one folder
