+++ scripts/SimplexUI/menu/genericPlugins/showFalloffs.py
