+++ scripts/SimplexUI/menu/mayaPlugins/snapToNeutral.py
