+++ scripts/SimplexUI/menu/mayaPlugins/tweakMix.py
@@ -1,6 +1,6 @@
 import maya.cmds as cmds
 from ...Qt.QtWidgets import QAction
-from ...mayaInterface import disconnected
+from ...interface.mayaInterface import disconnected
 from ...items import Combo, ProgPair
 from ...interfaceModel import coerceIndexToType
 from functools import partial
